polus-cell-nuclei-segmentation/src/requirements.txt

torch==1.7.0
torchvision==0.2.1
numpy==1.19.1
tqdm==4.41.1
pandas==0.25.3
lightgbm==2.3.1
Keras==2.1.6
matplotlib==3.1.1
keras_resnet==0.2.0
scikit_image==0.15.0
imgaug==0.3.0
opencv_python_headless==4.1.2.30
scikit_learn==0.19.1
telegram_send==0.25
tensorboardX==2.0
Pillow==6.2.2
scipy==1.1.0
cupy_cuda101==8.2.0
imagecodecs==2020.5.30
tifffile==2020.7.4

